backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# LangChain + LangGraph
langchain==0.2.1
langchain-core==0.2.1
langchain-groq==0.1.4
langgraph==0.1.1

# Document parsing
PyMuPDF==1.24.3
python-docx==1.1.2

# Vector DB
qdrant-client==1.9.1

# PostgreSQL
asyncpg==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1

# Embeddings
sentence-transformers==3.0.0
numpy==1.26.4

# Utilities
python-jose[cryptography]==3.3.0
httpx==0.27.0
tenacity==8.3.0
structlog==24.1.0